pyproject.toml
[project]
name = "codex-agent"
version = "0.1.0"
description = "Local-first research agent architecture scaffolding"
authors = [{ name = "Codex" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110,<0.112",
    "uvicorn[standard]>=0.23,<0.30",
    "pydantic>=2.5,<3",
    "jinja2>=3.1,<4",
    "pyyaml>=6,<7"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8",
    "httpx>=0.25,<0.28"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
